requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
python-telegram-bot==21.10
python-decouple==3.8
djangorestframework==3.15.2
requests==2.32.5